refactor: make Container a read-only view over a store; friendly FSStore ctor
- Container now wraps one IStore and exposes only get(); load/loadFile/add
  stay on the store (a store is the loadable unit, Container is a read-only view)
- FSStore gains a friendly constructor (cwd/prefix/suffix/extensions/mergeFn/
  naming/read) and builds its NamingScheme internally
- remove the createStore factory and the Options type (options are FSStoreOptions)
- clear the release-as pin now that v1.0.0 has shipped
