+++ AGENTS.md
@@ -2,7 +2,7 @@
 
 # Confinity — Agent Guide
 
-Confinity is a small TypeScript library for loading configurations in the context of a multi-package application. The `createStore()` factory builds an `FSStore` that discovers config files across one or many directories, parses them (via [locter](https://github.com/tada5hi/locter)), and exposes their values through a dotted-path getter that deep-merges matches from every loaded file. `FSStore` (the filesystem loader) extends `Store` (the pure query/merge engine), with the file-name convention factored out into a `NamingScheme`. It is published as an **ESM-only** package and is currently in **beta / work-in-progress**.
+Confinity is a small TypeScript library for loading configurations in the context of a multi-package application. An `FSStore` (the filesystem store) discovers config files across one or many directories, parses them (via [locter](https://github.com/tada5hi/locter)), and exposes their values through a dotted-path getter that deep-merges matches from every loaded file. `FSStore` extends `Store` (the pure query/merge engine), with the file-name convention factored out into a `NamingScheme`; a `Container` wraps any `IStore` as a read-only `get` view to hand to consumers. It is published as an **ESM-only** package and is currently in **beta / work-in-progress**.
 
 ## Quick Reference
 
@@ -37,11 +37,11 @@ npm run lint           # eslint (flat config)  (lint:fix to autofix)
 
 The package exposes a single ESM entry point (`src/index.ts`) that re-exports from `module.ts`, `naming/`, `store/`, and `types.ts`:
 
-- `createStore(options?)` — factory that builds a filesystem-backed `FSStore`
-- `FSStore` / `Store` — the filesystem loader (`load`, `loadFile`) and the pure query/merge engine (`add`, `get`); `FSStore extends Store`
+- `Container` — a read-only view over a single `IStore` (`get` only; no `load`/`loadFile`/`add`)
+- `FSStore` / `Store` — the filesystem store (`load`, `loadFile`) and the pure query/merge engine (`add`, `get`); `FSStore extends Store`
 - `NamingScheme` — the prefix/suffix/extensions convention (`toPatterns`, `toName`)
-- `INamingScheme`, `IStore` — the contracts callers inject through (`Options.naming`, `Options.read`)
-- `Options`, `Element`, `MergeFn`, `Reader`, `StoreOptions`, `NamingOptions`, `FSStoreOptions` — supporting types
+- `INamingScheme`, `IStore` — the class-implemented contracts (`INamingScheme` injected via `FSStoreOptions.naming`; `IStore` is what `Container` wraps)
+- `Element`, `MergeFn`, `Reader`, `StoreOptions`, `NamingOptions`, `FSStoreOptions` — supporting types
 
 ## Detailed Guides
 
@@ -58,7 +58,7 @@ implementation diverged from the proposal:
 
 1. [Extract a pure `Resolver`](.agents/plans/001-resolver.md) — query + merge over `Element[]` (now `Store`).
 2. [Extract a `NamingScheme`](.agents/plans/002-naming-scheme.md) — the prefix/suffix/extension convention (both directions).
-3. [Split `Loader` (I/O) vs `Store` (pure)](.agents/plans/003-loader-store-split.md) — the full seam (superset of 1 + 2), shipped as `FSStore extends Store` + the `createStore` factory.
+3. [Split `Loader` (I/O) vs `Store` (pure)](.agents/plans/003-loader-store-split.md) — the full seam (superset of 1 + 2), shipped as `FSStore extends Store`, with `Container` reduced to a read-only view over one store.
 
 ## Commits, Issues & Pull Requests
