+++ release-please-config.json
@@ -4,8 +4,6 @@
     "include-v-in-tag": true,
     "bump-minor-pre-major": true,
     "packages": {
-        ".": {
-            "release-as": "1.0.0"
-        }
+        ".": {}
     }
 }
