+++ src/module.ts
@@ -5,55 +5,23 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { createMerger } from 'smob';
-import { NamingScheme } from './naming';
-import { FSStore } from './store';
-import type { INamingScheme } from './naming';
-import type { Options } from './types';
-
-function normalizeExtensions(input?: string[]) : string[] {
-    if (
-        input &&
-        input.length > 0
-    ) {
-        return input.map((extension) => {
-            if (extension.startsWith('.')) {
-                return extension.substring(1);
-            }
-
-            return extension;
-        });
-    }
-
-    return ['conf', 'js', 'mjs', 'cjs', 'ts', 'mts', 'yml', 'yaml'];
-}
+import type { IStore } from './store';
 
 /**
- * Build a filesystem-backed config store from a set of friendly options.
+ * A read-only view over a single {@see IStore}.
  *
- * Normalizes the extensions, resolves the working directory, and wires a
- * {@see NamingScheme} (unless a custom {@see INamingScheme} is provided via
- * `naming`) into an {@see FSStore}. The returned store discovers, loads and
- * merges config files and answers dotted-path lookups.
- *
- * @param options
+ * Wrap a store (e.g. an {@see FSStore} you have already loaded) to hand
+ * consumers dotted-path lookups without exposing the loading or mutation
+ * surface — `load`/`loadFile`/`add` stay on the store itself.
  */
-export function createStore(options: Options = {}) : FSStore {
-    const extensions = normalizeExtensions(options.extensions);
+export class Container {
+    protected readonly store : IStore;
 
-    const naming : INamingScheme = options.naming ?? new NamingScheme({
-        prefix: options.prefix,
-        suffix: options.suffix,
-        extensions,
-    });
+    constructor(store: IStore) {
+        this.store = store;
+    }
 
-    return new FSStore({
-        cwd: options.cwd || process.cwd(),
-        naming,
-        mergeFn: options.mergeFn ?? createMerger({
-            array: false,
-            inPlace: false,
-        }),
-        read: options.read,
-    });
+    get<T = any>(key: string | string[]) : T | undefined {
+        return this.store.get<T>(key);
+    }
 }
